Moved ImageInput dropzone component to its on file
